refactor: move SlotController to component-base
